+++ product_code/examreg_service/api/src/test/scala/de/upb/cs/uc4/examreg/DefaultTestExamRegs.scala
@@ -4,20 +4,19 @@ import de.upb.cs.uc4.examreg.model.{ ExaminationRegulation, Module }
 
 trait DefaultTestExamRegs {
 
-  val examReg0: ExaminationRegulation = ExaminationRegulation("Computer Science v3", true, Seq(
+  val examReg0: ExaminationRegulation = ExaminationRegulation("Mathematics Bachelor v3", active = true, Seq(
     Module("M.1275.01158", "Math 1"),
     Module("M.1275.56002", "Math 2"),
-    Module("M.1358.91583", "Complexity Theory")
   ))
 
-  val examReg1: ExaminationRegulation = ExaminationRegulation("Computer Science v4", true, Seq(
-    Module("M.1568.98158", "Math"),
-    Module("M.4186.23588", "Secure Software Engineering")
+  val examReg1: ExaminationRegulation = ExaminationRegulation("Mathematics Bachelor v4", active = true, Seq(
+    Module("M.1275.01158", "Math 1"),
+    Module("M.1275.56002", "Math 2"),
+    Module("M.1275.13337", "Math 3"),
   ))
 
-  val examReg2: ExaminationRegulation = ExaminationRegulation("Computer Science v4", false, Seq(
-    Module("M.1568.98158", "Math"),
-    Module("M.4186.23588", "Secure Software Engineering")
+  val examReg2: ExaminationRegulation = ExaminationRegulation("Mathematics Bachelor v2", active = true, Seq(
+    Module("M.1568.98158", "Math 69"),
   ))
 
 }